requirements.txt
oauthlib==3.2.2
Pillow==10.3.0
Requests==2.32.3
requests_oauthlib==1.3.1
streamlit==1.34.0